updating taskmaster daemon
