--- shell.log
@@ -1,5 +1,0 @@
-2025-02-16 19:32:55,104 - ERROR - [ERROR] Unknown command: ls. Type 'help' to see the command list.
-2025-02-16 19:33:50,458 - INFO - [INFO] Starting program: hello
-2025-02-16 19:34:13,305 - INFO - [INFO] Starting program: hello my name is andrea
-2025-02-16 19:34:32,518 - INFO - 
-[INFO] Exiting shell...
