--- taskmaster.log
@@ -1,8 +1,0 @@
-2025-02-16 16:57:48,401 - ERROR - [ERROR] Unknown command: hello. Type 'help' to see the command list.
-2025-02-16 16:57:50,966 - ERROR - [ERROR] Unknown command: run hello. Type 'help' to see the command list.
-2025-02-16 16:57:54,086 - INFO - [INFO] Starting program: ls
-2025-02-16 16:58:32,382 - INFO - [INFO] Exiting shell...
-2025-02-16 17:54:26,828 - ERROR - [ERROR] Unknown command: soaeuthaosneh. Type 'help' to see the command list.
-2025-02-16 17:54:35,516 - INFO - [INFO] Starting program: ls
-2025-02-16 17:54:48,503 - INFO - 
-[INFO] Exiting shell...
